% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_nccluster_data.R
\name{fetch_new_NCtraining_data}
\alias{fetch_new_NCtraining_data}
\title{Generate data frame containing new customer data for entire year}
\usage{
fetch_new_NCtraining_data(year, conn)
}
\arguments{
\item{year}{is the year of interest. Must be in the format YYYY. The function
will convert to proper data type.}

\item{conn}{is the name of a valid data base connection.}
}
\value{
This function returns a single data frame.
}
\description{
\code{fetch_new_NCtraining_data} returns a data frame consisting of all new customers from the supplied year.
}
\details{
\strong{WARNING: THIS FUNCTION CAN TAKE >90 MINUTES TO COMPLETE QUERY}.

This function will produce a data frame containing all unique customers of a
given year, and >85 variables related to those customers. The query as
written is restricted to new customers and does not generate data beyond
information regarding their first purchase.
}
\examples{
\dontrun{
  con = dbConnect(odbc::odbc(), 'EDW_BA_PROD', uid = 'danwal1', pwd = 'XXXXXXXX')
}

 df = fetch_new_NCtraining_data(2018, conn = con)



}
